+++ polugins_type_gen/src/polugins_type_gen/cli.py
@@ -1,11 +1,17 @@
+import argparse
 import ast
 import importlib.resources as importlib_resources
-import sys
 from pathlib import Path
 
 from polugins.main import _get_namespaces  # TODO: makes this non private
+from typing_extensions import Optional
 
-from polugins_type_gen._ast import parse_from_current_env as _parse_from_current_env
+from polugins_type_gen._ast import (
+    parse_from_current_env as _parse_from_current_env,
+)
+from polugins_type_gen._ast import (
+    parse_from_venv as _parse_from_venv,
+)
 
 
 class NoNamespaceRegisteredException(Exception):
@@ -17,8 +23,8 @@ def has_version(version: str) -> bool:
     return (files / "_stubs" / version).is_dir()
 
 
-def create_stubs():
-    output_dir = Path("typings")
+def create_stubs(venv_path: Optional[Path] = None, output_dir: Optional[Path] = None):
+    output_dir = output_dir or Path("typings")
     all_namespaces = _get_namespaces()
 
     if all(namespace == {} for namespace in all_namespaces.values()):
@@ -27,7 +33,13 @@ def create_stubs():
 
     for extension_class, namespaces in all_namespaces.items():
         if namespaces:
-            stub_ast = _parse_from_current_env(extension_class)
+            print(f"Found namespaces for {extension_class.value}:")
+            for name, namespace in namespaces.items():
+                print(f"  {name}: {namespace}")
+            if venv_path:
+                stub_ast = _parse_from_venv(venv_path, extension_class)
+            else:
+                stub_ast = _parse_from_current_env(extension_class)
             new_class_nodes = []
             module_imports = set()
             for node in stub_ast.body:
@@ -54,18 +66,51 @@ def create_stubs():
             output_path = output_dir / extension_class.import_path
             output_path.parent.mkdir(exist_ok=True, parents=True)
             output_path.with_suffix(".pyi").write_text(ast.unparse(stub_ast))
+            print(f"Generated stubs for {extension_class.value} in {output_path.with_suffix('.pyi')}")
+
+
+def get_argparser():
+    parser = argparse.ArgumentParser(
+        prog="Polugins Type Gen",
+        description="Generates type stubs for polars with added namespaces.",
+    )
+    parser.add_argument(
+        "command",
+        help="The command to run",
+        choices=["stubs", "version"],
+    )
+    parser.add_argument(
+        "-i",
+        "--input",
+        help="The input venv to find Polars in. If not selected, uses currently activated venv.",
+        required=False,
+    )
+    parser.add_argument(
+        "-o",
+        "--output",
+        help="The output directory for the generated stubs. Default is ./typings/",
+        default="typings",
+        required=False,
+    )
+    return parser
 
 
 def cli():
-    if len(sys.argv) == 1:
-        from polugins_type_gen._version import __version__
+    parser = get_argparser()
+    args = parser.parse_args()
+    match args.command:
+        case "stubs":
+            create_stubs(
+                venv_path=Path(args.input) if args.input else None,
+                output_dir=Path(args.output) if args.output else None,
+            )
+        case "version":
+            from polugins_type_gen._version import __version__
 
-        print(f"Polugins Type Gen version: {__version__}")
-    elif sys.argv[1] == "stubs":
-        print("generating stubs at ./typings/")
-        create_stubs()
-    else:
-        print("Use `polugins stubs` to generate type stubs.")
+            print(f"Polugins Type Gen version: {__version__}")
+        case _:
+            msg = "Unknown command. Use `polugins stubs` to generate type stubs."
+            raise ValueError(msg)
 
 
 if __name__ == "__main__":